Time: 3 ms (38.5%), Space: 8.3 MB (17.35%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -4,7 +4,7 @@ public:
         int s = 0,e = x,ans = -1;
         while(s<=e){
             int m = s + (e-s)/2;
-            if((long)m*m <= x){
+            if((long long)m*m <= x){
                 ans = m;
                 s = m+1;
             }
